# docs/swagger.yaml
definitions:
  Error:
    properties:
      detail:
        type: string
    type: object
  Follow:
    properties:
      createdAt:
        type: string
      followedId:
        type: string
      followerId:
        type: string
      id:
        type: string
      updatedAt:
        type: string
    required:
    - followedId
    type: object
  FollowCreate:
    properties:
      followedID:
        type: string
    required:
    - followedID
    type: object
  Msg:
    properties:
      msg:
        type: string
    type: object
...
  ResetPassword:
    properties:
      newPassword:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - token
    type: object
  Token:
    properties:
      accessToken:
        type: string
      tokenType:
        type: string
    type: object
  User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      password:
        minLength: 8
        type: string
      updatedAt:
        type: string
    required:
    - email
    - password
    type: object
...
  ValidationError:
    properties:
      detail:
        additionalProperties: true
        type: object
    type: object
info:
  contact: {}
  title: Start
  version: 0.1.0
paths:
  /api/v1/account/current:
    get:
      consumes:
      - application/json
      description: Read current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Current
      tags:
      - Account
  /api/v1/account/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Login
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Token'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      summary: Login
      tags:
      - Account
  /api/v1/account/recover-password/{email}:
    post:
      consumes:
      - application/json
      description: Recover password
      parameters:
      - description: Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Msg'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      summary: Recover Password
      tags:
      - Account
  /api/v1/account/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ResetPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Msg'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      summary: Reset Password
      tags:
      - Account
  /api/v1/follows:
    get:
      consumes:
      - application/json
      description: Read follows
      parameters:
      - description: Follower Id
        in: query
        name: followerId
        type: string
      - description: Followed Id
        in: query
        name: followedId
        type: string
      - default: 0
        description: Skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Follow'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read Follows
      tags:
      - Follows
    post:
      consumes:
      - application/json
      description: Create follow
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/FollowCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Follow'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Create Follow
      tags:
      - Follows
  /api/v1/follows/{follow_id}:
    delete:
      consumes:
      - application/json
      description: Delete follow
      parameters:
      - description: Follow Id
        in: path
        name: follow_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Msg'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Delete Follow
      tags:
      - Follows
    get:
      consumes:
      - application/json
      description: Read follow
      parameters:
      - description: Follow Id
        in: path
        name: follow_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Follow'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ValidationError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
      security:
      - ApiKeyAuth: []
      summary: Read Follow
      tags:
      - Follows
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"